requirements.txt
# server
Django
django-environ==0.10.0
psycopg2-binary==2.9.9
django-cors-headers
dj-stripe
djangorestframework
gunicorn
google-api-python-client
google-cloud-secret-manager
google-cloud-logging
google-cloud-storage
django-storages[google]
av
io

# ml
# torch-cpu (for ultralytics)
https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp311-cp311-linux_x86_64.whl
# https://download.pytorch.org/whl/cpu/torchvision-0.15.0%2Bcpu-cp311-cp311-linux_x86_64.whl
# tensorflow-cpu

numpy
opencv-python-headless
matplotlib
scipy
ultralytics
lapx
openai
pytube